imp: exec: avoid invoking kill(2) with pid <= 0
Problem: The fwd_signal() function in src/imp/exec/exec.c doesn't check
its argument for validity before invoking kill(2). This could lead
to the IMP process killing itself, its process group, or all processes.

Check for imp_child > 0 for good measure before invoking kill().

+++ src/imp/exec/exec.c
@@ -207,7 +207,8 @@ static void __attribute__((noreturn)) imp_exec (struct imp_exec *exec)
 
 static void fwd_signal (int signal)
 {
-    kill (imp_child, signal);
+    if (imp_child > 0)
+        kill (imp_child, signal);
 }
 
 /*  Setup signal handlers in the IMP for common signals which
